% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps_doc.R
\name{bps_doc}
\alias{bps_doc}
\title{bps_doc}
\usage{
bps_doc(geography = "place")
}
\arguments{
\item{geography}{Character, length 1. One of "state", "county", "msa", or
"place".}
}
\description{
Source: https://www2.census.gov/econ/bps/Documentation/
}
\details{
Open BPS documentation (PDF) in system default PDF viewer
}
