% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bps_read.R
\name{bps_read}
\alias{bps_read}
\title{Read annual BPS data}
\usage{
bps_read(file, repair_names = F)
}
\arguments{
\item{file}{Character, length 1. Path to BPS annual data file.}

\item{repair_names}{Logical, length 1. If true, applies a custom name repair
function. Does NOT use the 'tibble' package's name repair method.}
}
\value{
Data frame. At the moment, the 'tibble' package is still maturing.
Its name repair method has changed quite a bit in the past month or so.
Once it's stable, perhaps I'll use it.
}
\description{
Load BPS data and clean messy column names
}
